Add files via upload
Pluto notebook, mainly containing the function "boundaryplot3d" which can plot the characteristics of a 3d simplexgrid builder. It is Makie based. Only triangular faces are supported yet. The function "boundaryplot3d" uses the function "give_builder_vertsfaces" which returns the vertices and faces stored in the builder in a useful format, the function is not optimized yet.
Since the function is supposed to get integrated in the SimplexGridFactory package a number of functions from this package are used, some of them are not exported though, thus they are copied from the source code.
Something similar is done with the "gridplot33d" function (it is for 3d...). Its source code is taken from the "GridVisualize.gridplot!" function with a few modifications for the pluto environment where angles can be controlled via sliders.